NZBGet: Attribution update to HashRenamer.py script.
Attribution was given to 
[Clinton-Hall](https://github.com/clinton-hall/GetScripts) since the 
base template of their NZBGet scripts was used. But despite what 
copy-cat groups think, 90-99 percent of the script was built by us. But 
unlike them, we dont copy entire scripts from others and rename it to 
obfuscate from the origin source - that's their modus operandi, not 
ours. ;)